+++ Changelog.md
@@ -27,6 +27,7 @@
   - [#210](../../issues/210) - Add dedicated "Diagnostics" view
   - [#212](../../issues/212) - Add binding diagnostics
   - [#213](../../issues/213) - Add DynamicResource diagnostics
+  - [#219](../../issues/219) - Add a warning to zoomer if target has TextOptions.TextFormattingMode=Display
   - [#226](../../issues/226) - Add support for ARM/ARM64
   - [#227](../../issues/227) - Add support for .NET 6